Preview Form доработка
